requirements.txt
streamlit
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.1
seaborn==0.12.2
tensorflow==2.13.0
torch==2.0.1
transformers==4.33.0
datasets==2.14.4
nltk==3.8.1
spacy==3.6.1
imbalanced-learn==0.11.0
streamlit-option-menu==0.3.6
huggingface-hub==0.16.4
typing_extensions
python-dotenv